scripts/upgrade-verify/requirements.txt
requests
websocket-client
cryptography
pytest
pytest-html
pytest-metadata
